DCNDL、NDL-SRUのインポート処理の実装（途中） - SRUから取得後の実装
