+++ configservice.go
@@ -62,6 +62,10 @@ const OKLBL = "\u001B[32mOK\u001B[0m"
 const TIKAAPPURL = "https://dlcdn.apache.org/tika/2.4.1/tika-server-standard-2.4.1.jar"
 const SOLRAPPURL = "https://www.apache.org/dyn/closer.lua/lucene/solr/8.11.2/solr-8.11.2.zip?action=download"
 
+var (
+	KCONF *KENVCONF
+)
+
 func DownloadApps() {
 	log.Info().Msg("download tika-server")
 
@@ -469,6 +473,7 @@ func dism(msg string, isError bool) {
 // display mode : full or error-only
 func CheckConfigAndConnections(displaymode string) error {
 
+	var last_error error
 	DisplayModeOnCheckFunction = displaymode
 
 	fmt.Printf("check start: %s\n", DisplayModeOnCheckFunction)
@@ -491,6 +496,7 @@ func CheckConfigAndConnections(displaymode string) error {
 
 	f, err := os.Stat(extdirname)
 	if err == nil && f.IsDir() {
+		last_error = err
 		dism(fmt.Sprintf("ext dirname:"+OKLBL+" (%s)\n", extdirname), MSGINFO)
 	} else {
 		dism(fmt.Sprintf("ext dirname:"+NGLBL+" (%s)\n", extdirname), MSGERROR)
@@ -499,6 +505,7 @@ func CheckConfigAndConnections(displaymode string) error {
 	// step3 : check java
 	out, err := exec.Command("java", "--version").Output()
 	if err != nil {
+		last_error = err
 		dism("Java command:"+NGLBL+"\n", MSGERROR)
 		dism(err.Error(), MSGERROR)
 	} else {
@@ -511,11 +518,16 @@ func CheckConfigAndConnections(displaymode string) error {
 	// step3-b : check java home
 	javahome := os.Getenv("JAVA_HOME")
 	if javahome == "" {
+		last_error = fmt.Errorf("no set JAVA_HOME")
 		dism("Env JAVA_HOME:"+NGLBL+"\n", MSGERROR)
 	} else {
 		javaexe := filepath.Join(javahome, "bin", "java.exe")
+		if runtime.GOOS != "windows" {
+			javaexe = filepath.Join(javahome, "bin", "java")
+		}
 		_, err := os.Stat(javaexe)
 		if err != nil {
+			last_error = fmt.Errorf("no find java.exe")
 			dism(fmt.Sprintf("ERROR: java.exe "+NGLBL+" not found in %s : Please set JAVA_HOME to a valid JRE / JDK directory.\n", javaexe), MSGERROR)
 		}
 	}
@@ -523,6 +535,7 @@ func CheckConfigAndConnections(displaymode string) error {
 	// step4 : check solr
 	solrhome := filepath.Join(cfg.Solr.Home)
 	if f, err := os.Stat(solrhome); os.IsNotExist(err) || !f.IsDir() {
+		last_error = fmt.Errorf("no exists solr home: %s", solrhome)
 		dism(fmt.Sprintf("Solr home:"+NGLBL+" %s\n", solrhome), MSGERROR)
 	} else {
 		dism(fmt.Sprintf("Solr home:"+OKLBL+" %s\n", solrhome), MSGINFO)
@@ -530,6 +543,7 @@ func CheckConfigAndConnections(displaymode string) error {
 
 	solrbin := filepath.Join(cfg.Solr.Home, "bin")
 	if f, err := os.Stat(solrbin); os.IsNotExist(err) || !f.IsDir() {
+		last_error = fmt.Errorf("no exists solr bin: %s", solrbin)
 		dism(fmt.Sprintf("Solr bin:"+NGLBL+" %s\n", solrbin), MSGERROR)
 	} else {
 		dism(fmt.Sprintf("Solr bin:"+OKLBL+" %s\n", solrbin), MSGINFO)
@@ -537,6 +551,7 @@ func CheckConfigAndConnections(displaymode string) error {
 
 	defaultsolrconfigset := filepath.Join(cfg.Solr.Home, "server", "solr", "configsets", "_default")
 	if f, err := os.Stat(defaultsolrconfigset); os.IsNotExist(err) || !f.IsDir() {
+		last_error = fmt.Errorf("no default config set: %s", defaultsolrconfigset)
 		dism(fmt.Sprintf("Solr default config set:"+NGLBL+" %s\n", defaultsolrconfigset), MSGERROR)
 	} else {
 		dism(fmt.Sprintf("Solr default config set:"+OKLBL+" %s\n", defaultsolrconfigset), MSGINFO)
@@ -546,36 +561,40 @@ func CheckConfigAndConnections(displaymode string) error {
 	uri := cfg.Solr.Serveruri + "/solr"
 	si, err := solr.NewSolrInterface(uri, cfg.Solr.Corename)
 	if err != nil {
+		last_error = fmt.Errorf("no solr coneection")
 		dism(fmt.Sprintf("Solr connection:"+NGLBL+" error (1) Path:%s %s\n", cfg.Solr.Serveruri, cfg.Solr.Corename), MSGERROR)
 	} else {
 		//admin core
 		vs, err := SolrServerPing(cfg.Solr.Serveruri)
 		if err != nil {
+			last_error = fmt.Errorf("no solr admincore")
 			dism(fmt.Sprintf("Solr admincore:"+NGLBL+" error: %s\n", err.Error()), MSGERROR)
 		} else {
 			dism(fmt.Sprintf("Solr admincore:"+OKLBL+" solr-spec-version:%s\n", vs), MSGINFO)
-		}
 
-		//core check
-		_, qtime, err := si.Ping()
-		if err != nil {
-			dism(fmt.Sprintf("Solr core ping:"+NGLBL+" (%s %s)\n", cfg.Solr.Serveruri, cfg.Solr.Corename), MSGERROR)
-		} else {
-			dism(fmt.Sprintf("Solr core ping:"+OKLBL+" (%s %s) qtime:%d\n", cfg.Solr.Serveruri, cfg.Solr.Corename, qtime), MSGINFO)
+			//core check
+			_, qtime, err := si.Ping()
+			if err != nil {
+				last_error = fmt.Errorf("solr ping error")
+				dism(fmt.Sprintf("Solr core ping:"+NGLBL+" (%s %s)\n", cfg.Solr.Serveruri, cfg.Solr.Corename), MSGERROR)
+			} else {
+				dism(fmt.Sprintf("Solr core ping:"+OKLBL+" (%s %s) qtime:%d\n", cfg.Solr.Serveruri, cfg.Solr.Corename, qtime), MSGINFO)
 
-			//TODO: core and schema check
+				//TODO: core and schema check
+			}
 		}
 	}
 
 	// step6 : check tika
 	vs, err := TikaPing(cfg.Tika.Serveruri)
 	if err != nil {
+		last_error = fmt.Errorf("tika ping error")
 		dism(fmt.Sprintf("Tika ping:"+NGLBL+" (%s) %s\n", cfg.Tika.Serveruri, err), MSGERROR)
 	} else {
 		dism(fmt.Sprintf("Tika ping:"+OKLBL+" (%s) %s\n", cfg.Tika.Serveruri, vs), MSGINFO)
 	}
 
-	return nil
+	return last_error
 }
 
 func TikaPing(tikauri string) (string, error) {
