+++ testdata/isbn.txt
@@ -1,15 +1,11 @@
 123
 9784805207505
-978-4-80-520750-5
-9784480689108
+978-4-48-068910-8
 9784120042591
 9784087472226
-9784003390214
-9784150766023
-9784585205012
 9784864102537
-9784864102537
-
-
+9784864102538
 
 9784774189307
+
+
